+++ src/main/kotlin/dev/aaronhowser/mods/genetics_resequenced/attachment/GeneCooldowns.kt
@@ -80,7 +80,7 @@ class GeneCooldowns(
 				return false
 			}
 
-			val newEntry = Entry(gene, duration, notify)
+			val newEntry = Entry(gene, event.cooldownTicks, notify)
 			newEntry.notifyStart(entity)
 
 			val newCooldowns = entity.geneCooldowns.cooldowns + newEntry