+++ src/main/kotlin/dev/aaronhowser/mods/genetics_resequenced/attachment/TemporaryGenesData.kt
@@ -11,15 +11,18 @@ import dev.aaronhowser.mods.genetics_resequenced.event.custom.TemporaryGeneAdded
 import dev.aaronhowser.mods.genetics_resequenced.event.custom.TemporaryGeneRemovedEvent
 import dev.aaronhowser.mods.genetics_resequenced.gene.Gene
 import dev.aaronhowser.mods.genetics_resequenced.gene.Gene.Companion.getName
+import dev.aaronhowser.mods.genetics_resequenced.gene.Gene.Companion.isDisabled
 import dev.aaronhowser.mods.genetics_resequenced.gene.Gene.Companion.isGene
 import dev.aaronhowser.mods.genetics_resequenced.gene.Gene.Companion.isHelixOnly
 import dev.aaronhowser.mods.genetics_resequenced.gene.behavior.TickGenes
 import dev.aaronhowser.mods.genetics_resequenced.registry.ModAttachmentTypes
+import dev.aaronhowser.mods.genetics_resequenced.registry.ModGenes
 import net.minecraft.core.Holder
 import net.minecraft.network.RegistryFriendlyByteBuf
 import net.minecraft.network.chat.Component
 import net.minecraft.network.codec.ByteBufCodecs
 import net.minecraft.network.codec.StreamCodec
+import net.minecraft.resources.ResourceKey
 import net.minecraft.world.entity.EntityType
 import net.minecraft.world.entity.LivingEntity
 import thedarkcolour.kotlinforforge.neoforge.forge.FORGE_BUS
@@ -60,6 +63,17 @@ data class TemporaryGenesData(
 		val LivingEntity.temporaryGeneHolders: List<Holder<Gene>>
 			get() = this.temporaryGenes.map(TemporaryGene::geneHolder)
 
+		@JvmStatic
+		fun LivingEntity.hasTemporaryGene(geneHolder: Holder<Gene>): Boolean {
+			return !geneHolder.isDisabled && this.temporaryGeneHolders.any { it.isGene(geneHolder) }
+		}
+
+		@JvmStatic
+		fun LivingEntity.hasTemporaryGene(geneRk: ResourceKey<Gene>): Boolean {
+			val geneHolder = ModGenes.fromResourceKey(registryAccess(), geneRk) ?: return false
+			return hasTemporaryGene(geneHolder)
+		}
+
 		fun tickTemporaryGenes(entity: LivingEntity) {
 			val toRemove = mutableListOf<TemporaryGene>()
 			val tempGenes = entity.temporaryGenes
@@ -75,21 +89,47 @@ data class TemporaryGenesData(
 			}
 		}
 
+		@JvmStatic
+		fun LivingEntity.removeTemporaryGene(
+			geneRk: ResourceKey<Gene>
+		): Boolean {
+			val geneHolder = ModGenes.fromResourceKey(registryAccess(), geneRk) ?: return false
+			return removeTemporaryGene(geneHolder)
+		}
+
+		@JvmStatic
 		fun LivingEntity.removeTemporaryGene(
 			geneHolderToRemove: Holder<Gene>
-		) {
+		): Boolean {
 			val existingList = this.temporaryGenes.toMutableList()
 			val wasRemoved = existingList.removeIf { it.geneHolder.isGene(geneHolderToRemove) }
-			if (!wasRemoved) return
+			if (!wasRemoved) return false
+
+			val eventPre = TemporaryGeneRemovedEvent.Pre(this, geneHolderToRemove)
+			if (FORGE_BUS.post(eventPre).isCanceled) {
+				GeneticsResequenced.LOGGER.debug("Event was canceled: $eventPre")
+				return false
+			}
 
 			if (geneHolderToRemove.value().potions.isNotEmpty()) {
 				TickGenes.handlePotionGeneRemoved(this, geneHolderToRemove)
 			}
 
-			val event = TemporaryGeneRemovedEvent(this, geneHolderToRemove)
-			FORGE_BUS.post(event)
-
 			this.temporaryGenes = existingList
+
+			val eventPost = TemporaryGeneRemovedEvent.Post(this, geneHolderToRemove)
+			FORGE_BUS.post(eventPost)
+
+			return true
+		}
+
+		@JvmStatic
+		fun LivingEntity.addTemporaryGene(
+			newGeneRk: ResourceKey<Gene>,
+			durationTicks: Int
+		): Boolean {
+			val geneHolder = ModGenes.fromResourceKey(registryAccess(), newGeneRk) ?: return false
+			return addTemporaryGene(geneHolder, durationTicks)
 		}
 
 		@JvmStatic
@@ -136,19 +176,20 @@ data class TemporaryGenesData(
 				GeneticsResequenced.LOGGER.debug("Event was canceled: $eventPre")
 				return false
 			}
+			val finalDurationTicks = eventPre.durationTicks
 
 			val existingList = this.temporaryGenes.toMutableList()
 
 			val existingTempGene = existingList.find { it.geneHolder.isGene(newGeneHolder) }
 			if (existingTempGene != null) {
-				existingTempGene.ticksRemaining = durationTicks
+				existingTempGene.ticksRemaining = finalDurationTicks
 			} else {
-				existingList.add(TemporaryGene(newGeneHolder, durationTicks))
+				existingList.add(TemporaryGene(newGeneHolder, finalDurationTicks))
 			}
 
 			this.temporaryGenes = existingList
 
-			val eventPost = TemporaryGeneAddedEvent.Post(this, newGeneHolder, durationTicks)
+			val eventPost = TemporaryGeneAddedEvent.Post(this, newGeneHolder, finalDurationTicks)
 			FORGE_BUS.post(eventPost)
 
 			return true