+++ src/main/kotlin/dev/aaronhowser/mods/genetics_resequenced/compatibility/kubejs/kube_event/GeneKubeEvent.kt
@@ -1,0 +1,9 @@
+package dev.aaronhowser.mods.genetics_resequenced.compatibility.kubejs.kube_event
+
+import dev.aaronhowser.mods.genetics_resequenced.gene.Gene
+import dev.latvian.mods.kubejs.entity.KubeLivingEntityEvent
+import net.minecraft.core.Holder
+
+interface GeneKubeEvent : KubeLivingEntityEvent {
+	fun getGene(): Holder<Gene>
+}